Phase 2 move files.
