+++ libshortfin/src/shortfin/CMakeLists.txt
@@ -8,4 +8,3 @@ add_subdirectory(array)
 add_subdirectory(local)
 add_subdirectory(process)
 add_subdirectory(support)
-add_subdirectory(systems)
