+++ libshortfin/src/shortfin/local/systems/amdgpu.cc
@@ -4,7 +4,7 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "shortfin/systems/amdgpu.h"
+#include "shortfin/local/systems/amdgpu.h"
 
 #include "shortfin/support/logging.h"
